+++ front/components/common/Footer.js
