--- front/components/common/Logo.js
@@ -1,20 +1,0 @@
-import React from "react";
-
-export default function Logo() {
-  return (
-    <>
-      <svg height="26" viewBox="0 0 44 39" fill="none" xmlns="http://www.w3.org/2000/svg">
-        <path d="M17.5 20.4286H21.7286V39H17.5V20.4286Z" fill="#58AFFF"/>
-        <path d="M26 13L21.7286 16.12V39V39C24.0876 39 26 37.0876 26 34.7285V13Z" fill="#4A91F5"/>
-        <path d="M21.7286 10.8643C21.7286 4.8641 26.5927 0 32.5929 0V0V8.35714L21.7286 16.7143V10.8643Z" fill="#58AFFF"/>
-        <path d="M43.4571 8.35715C43.4571 12.9727 39.7155 16.7143 35.1 16.7143H21.7286L32.5928 8.35715H43.4571V8.35715Z" fill="#4A91F5"/>
-        <path d="M32.5928 0H43.4571L32.5928 8.35714V0Z" fill="#66D1FF"/>
-        <path d="M43.4571 0V8.35714H32.5928L43.4571 0Z" fill="#58C2FD"/>
-        <path d="M21.7286 14.5785C21.7286 8.57836 16.8645 3.71426 10.8643 3.71426V3.71426V12.0714L21.7286 20.4285V14.5785Z" fill="#58AFFF"/>
-        <path d="M7.62939e-06 12.0714C7.62939e-06 16.6869 3.74163 20.4286 8.35715 20.4286H21.7286L10.8643 12.0714H7.62939e-06V12.0714Z" fill="#4A91F5"/>
-        <path d="M10.8643 3.71426H-1.19209e-06L10.8643 12.0714V3.71426Z" fill="#58C2FD"/>
-        <path d="M-1.19209e-06 3.71426V12.0714H10.8643L-1.19209e-06 3.71426Z" fill="#3CB7FC"/>
-      </svg>
-    </>
-  );
-}
