+++ front/components/common/LogoWhite.module.css
@@ -1,0 +1,7 @@
+.logo {
+  display: inline;
+}
+
+.logo path {
+  fill: white;
+}
