+++ front/components/common/TopButton.js
@@ -2,19 +2,28 @@ import React from "react";
 import styles from "./TopButton.module.css";
 
 const TopButton = () => {
-
   return (
     <div>
       <a href="#">
-        <div className={styles.button} tabIndex="0" role="button" aria-label="맨 위로" title="맨 위로">
+        <div
+          className={styles.button}
+          tabIndex="0"
+          role="button"
+          aria-label="맨 위로"
+          title="맨 위로"
+        >
           <div className={styles.icon}>
-              <img src={require('../src/image/ToTop.png')} width="16" height="24" />
+            <img
+              src={require("../../src//image/ToTop.png")}
+              width="16"
+              height="24"
+            />
           </div>
           <div className={styles.txt}>맨 위로</div>
         </div>
       </a>
     </div>
   );
-}
+};
 
 export default TopButton;
